requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
ndg-httpsclient==0.5.1
oauthlib==3.2.2
pillow==11.2.1
pusher==3.3.3
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
